requirements.txt
python-telegram-bot==20.8
openai
python-dotenv

